change some sql creation code for better grouping moved all module efffects code into it's own testing directory started dungeon manager code - dungeon data manager coded and working - code started for saving/loading persistant dungeons (untested)
added many notes, and begun code for (finally) fixing orbit

+++ src/eve-server/planet/Planet.cpp
@@ -38,7 +38,8 @@ void PlanetDataMgr::_Populate()
 
     m_db->GetPlanetData(*res);
     while (res->GetRow(row)) {
-        m_planetData.insert(std::pair<uint32, uint16>(row.GetInt(0), row.GetInt(1)));
+        // SELECT planet.typeID, resource.typeID
+        m_planetData.insert(std::pair<uint32, uint32>(row.GetInt(0), row.GetInt(1)));
     }
 
     //cleanup
